fix(kafka): use double instead of int as backoff multiplier

+++ sda-commons-starter-kafka/src/main/java/org/sdase/commons/spring/boot/kafka/config/KafkaConsumerConfig.java
@@ -16,7 +16,7 @@ public record KafkaConsumerConfig(@NotNull RetryConfig retry, DLTConfig dlt) {
       @NotNull Integer maxRetries,
       @NotNull Long initialBackoffInterval,
       @NotNull Long maxBackoffInterval,
-      @NotNull Long backoffMultiplier) {}
+      @NotNull Double backoffMultiplier) {}
 
   public record DLTConfig(String pattern) {}
 }
